SNEK-20: .) created gameboard-view.fxml and GameboardViewController.java .) renamed SnakeController.java to MainViewController.java .) changed Button in main-view.fxml to show text "Start Game!" instead of "Click Me"; also changed it's functionality to set the scene gameScreen
