+++ src/main/java/at/ac/fhcampuswien/snake/SnakeApp.java
@@ -1,6 +1,7 @@
 package at.ac.fhcampuswien.snake;
 
-import at.ac.fhcampuswien.snake.controller.SnakeController;
+import at.ac.fhcampuswien.snake.controller.GameboardViewController;
+import at.ac.fhcampuswien.snake.controller.MainViewController;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -19,21 +20,14 @@ public class SnakeApp extends Application {
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(SnakeApp.class.getResource("main-view.fxml"));
         Scene startScreen = new Scene(fxmlLoader.load(), SCREEN_SIZE_MEDIUM, SCREEN_SIZE_MEDIUM);
+        MainViewController mainViewController = fxmlLoader.getController();
+        mainViewController.setStage(stage);
+
 
         startScreen.getStylesheets().add("css/app.css");
-
-        SnakeController controller = fxmlLoader.getController();
-        ImageView logo = controller.getLogo();
+        ImageView logo = mainViewController.getLogo();
 
         logo.fitWidthProperty().bind(stage.widthProperty());
-
-        Pane gameBoardPane = controller.getGameBoard();
-        GameBoard gameBoard = new GameBoard(gameBoardPane);
-        gameBoard.startGame();
-
-        stage.setOnCloseRequest(event -> {
-            gameBoard.stopGame();
-        });
 
         stage.setTitle(TITLE);
         stage.setScene(startScreen);
