+++ src/main/resources/at/ac/fhcampuswien/snake/gameboard-view.fxml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<VBox xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="at.ac.fhcampuswien.snake.controller.GameboardViewController"
+            prefHeight="400.0" prefWidth="600.0">
+    <VBox>
+        <MenuBar>
+            <Menu text="Menu">
+            </Menu>
+        </MenuBar>
+        <Pane fx:id="gameBoard">
+
+        </Pane>
+        <Pane fx:id="scoreBoard">
+
+        </Pane>
+    </VBox>
+
+</VBox>
